+++ paper_uploads/admin/collection.py
@@ -26,22 +26,22 @@ class CollectionAdminBase(UploadedFileBase):
                 name="%s_%s_delete" % info,
             ),
             path(
-                "upload_item/",
+                "upload-item/",
                 self.admin_site.admin_view(self.upload_item_view_class.as_view()),
                 name="%s_%s_upload_item" % info,
             ),
             path(
-                "delete_item/",
+                "delete-item/",
                 self.admin_site.admin_view(self.delete_item_view_class.as_view()),
                 name="%s_%s_delete_item" % info,
             ),
             path(
-                "change_item/",
+                "change-item/",
                 self.admin_site.admin_view(self.change_item_view_class.as_view()),
                 name="%s_%s_change_item" % info,
             ),
             path(
-                "sort_items/",
+                "sort-items/",
                 self.admin_site.admin_view(self.sort_items_view_class.as_view()),
                 name="%s_%s_sort_items" % info,
             ),